dat-service-extension: removed DynamicAttributeTokenProvider interface
